+++ src/Company/CompanyService.php
@@ -24,18 +24,20 @@ use Wizaplace\SDK\Exception\NotFound;
 final class CompanyService extends AbstractService
 {
     /**
-     * @param CompanyRegistration $companyRegistration
+     * @param AbstractCompanyRegistration $companyRegistration
      *
      * @return CompanyRegistrationResult
      * @throws AuthenticationRequired
      * @throws \GuzzleHttp\Exception\GuzzleException
      * @throws \Wizaplace\SDK\Exception\JsonDecodingError
      */
-    public function register(CompanyRegistration $companyRegistration): CompanyRegistrationResult
+    public function register(AbstractCompanyRegistration $companyRegistration): CompanyRegistrationResult
     {
         $this->client->mustBeAuthenticated();
 
-        $responseData = $this->client->post('companies', [
+        $endpoint = $companyRegistration->isC2C() ? 'companies/c2c' : 'companies';
+
+        $responseData = $this->client->post($endpoint, [
             RequestOptions::JSON => [
                 'name' => $companyRegistration->getName(),
                 'email' => $companyRegistration->getEmail(),
@@ -70,18 +72,14 @@ final class CompanyService extends AbstractService
     /**
      * Register a new C2C company (Customer-To-Customer, aka private individual).
      *
-     * @param string      $companyName
-     * @param string|null $iban
-     * @param string|null $bic
-     * @param array       $files
-     *
-     * @return CompanyRegistrationResult
      * @throws AuthenticationRequired
-     * @throws \GuzzleHttp\Exception\GuzzleException
-     * @throws \Wizaplace\SDK\Exception\JsonDecodingError
+     *
+     * @deprecated
      */
     public function registerC2CCompany($companyName = '', ?string $iban = null, ?string $bic = null, array $files = []): CompanyRegistrationResult
     {
+        @trigger_error('The method "registerC2CCompany" is deprecated, use "register" with CompanyC2CRegistration::class instead.', E_USER_DEPRECATED);
+
         $this->client->mustBeAuthenticated();
 
         $responseData = $this->client->post('companies/c2c', [